Fix evaluate_model: Convert targets to float for BCEWithLogitsLoss

+++ main_train.py
@@ -88,7 +88,7 @@ def evaluate_model(model, val_loader, criterion, device):
         for X, y in val_loader:
             X, y = X.to(device), y.to(device)
             outputs = model(X).squeeze()
-            loss = criterion(outputs, y)
+            loss = criterion(outputs, y.float())
             total_loss += loss.item()
             
             preds = (torch.sigmoid(outputs) > 0.5).cpu().numpy()
